x/inference/types/message_run_inference_test.go
package types

import (
	"testing"

	"cosmos-llm/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgRunInference_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgRunInference
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgRunInference{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgRunInference{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
